refactor,style: move shared styled components to global styles
