+++ src/components/Footer.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { StyledFooter } from '../styles/Footer.styled';
+import { FooterLink, StyledFooter } from '../styles/Footer.elements';
 
 export default function Footer() {
   return (
     <StyledFooter>
-      <a href="https://profy.dev/employers">profy.dev</a>
+      <FooterLink as="a" href="https://profy.dev/employers">profy.dev</FooterLink>
       <Link to="/"><img src="./images/compact-logo.svg" alt="compact" /></Link>
-      <Link to="terms">Terms & Privacy</Link>
+      <FooterLink to="terms">Terms & Privacy</FooterLink>
     </StyledFooter>
   );
 }
